change rate type to tinyint

+++ database/migrations/2021_05_22_093012_create_ratings_table.php
@@ -18,7 +18,7 @@ class CreateRatingsTable extends Migration
             $table->unsignedBigInteger('book_id');
             $table->unsignedBigInteger('user_id');
             $table->unique(['book_id', 'user_id']);
-            $table->unsignedInteger('rate');
+            $table->tinyInteger('rate');
             $table->softDeletes();
             $table->timestamps();
